LPD-55486 Fix NullPointerException caused by LPD-44069
Before LPD-44069, the test was using FileSystems.newFileSystem(Path path, ClassLoader loader).
However, with Java 17, the FileSystems API changed and LPD-44069 introduced the wrong version
of the method that expects a non-null Map object.

This should be replaced with the simpler version of the method introduced by Java 17 that only
requires a Path object.

See documentation here for more information:
- https://docs.oracle.com/javase/8/docs/api/java/nio/file/FileSystems.html
- https://docs.oracle.com/en/java/javase/17/docs/api/java.base/java/nio/file/FileSystems.html